add pghero

+++ config/initializers/sidekiq_cron.rb
@@ -5,14 +5,14 @@ return if Rails.application.secrets.redis[:url].blank?
 
 schedule_file = "config/schedule.yml"
 
+unless Rails.env.production? && Sidekiq.server?
+  Sidekiq.redis { |conn| conn.flushdb }
+end
+
 if Sidekiq.server?
   Sidekiq::Cron::Job.destroy_all!
 end
 
-if File.exist?(schedule_file) && Sidekiq.server? && Rails.env.production?
+if File.exist?(schedule_file) && Sidekiq.server?
   Sidekiq::Cron::Job.load_from_hash YAML.load_file(schedule_file)
-end
-
-unless Rails.env.production?
-  Sidekiq.redis { |conn| conn.flushdb }
 end
